Fix broken tests and quick enhancements

+++ db/migrate/20180315164256_add_tracker_type_to_spree_trackers.rb
@@ -1,9 +1,13 @@
-class AddTrackerTypeToSpreeTrackers < SolidusSupport::Migration[4.2][5.0]
+class AddTrackerTypeToSpreeTrackers < SolidusSupport::Migration[4.2]
   def up
-    add_column :spree_trackers, :tracker_type, :string
+    unless column_exists?(:spree_trackers, :tracker_type)
+      add_column :spree_trackers, :tracker_type, :string
+    end
   end
 
   def down
-    remove_column :spree_trackers, :tracker_type
+    if column_exists?(:spree_trackers, :tracker_type)
+      remove_column :spree_trackers, :tracker_type
+    end
   end
 end
